refactor(train): move pipeline MTP loss wrapper
